+++ config/laranail/authkit.php
@@ -32,41 +32,4 @@ return [
         ],
     ],
 
-    'social' => [
-        'google' => [
-            'client_id'     => env(key: 'AUTHKIT_GOOGLE_CLIENT_ID'),
-            'client_secret' => env(key: 'AUTHKIT_GOOGLE_CLIENT_SECRET'),
-            'redirect'      => env(key: 'AUTHKIT_GOOGLE_REDIRECT'),
-            'scopes'        => ['openid', 'profile', 'email'],
-        ],
-
-        'facebook' => [
-            'client_id'     => env(key: 'AUTHKIT_FACEBOOK_CLIENT_ID'),
-            'client_secret' => env(key: 'AUTHKIT_FACEBOOK_CLIENT_SECRET'),
-            'redirect'      => env(key: 'AUTHKIT_FACEBOOK_REDIRECT'),
-            'scopes'        => ['email', 'public_profile'],
-        ],
-
-        'twitter' => [
-            'client_id'     => env(key: 'AUTHKIT_TWITTER_CLIENT_ID'),
-            'client_secret' => env(key: 'AUTHKIT_TWITTER_CLIENT_SECRET'),
-            'redirect'      => env(key: 'AUTHKIT_TWITTER_REDIRECT'),
-            'scopes'        => ['users.read', 'users.email', 'tweet.read'],
-        ],
-
-        'linkedin' => [
-            'client_id'     => env(key: 'AUTHKIT_LINKEDIN_CLIENT_ID'),
-            'client_secret' => env(key: 'AUTHKIT_LINKEDIN_CLIENT_SECRET'),
-            'redirect'      => env(key: 'AUTHKIT_LINKEDIN_REDIRECT'),
-            'scopes'        => ['openid', 'profile', 'email'],
-        ],
-
-        'paypal' => [
-            'client_id'     => env(key: 'AUTHKIT_PAYPAL_CLIENT_ID'),
-            'client_secret' => env(key: 'AUTHKIT_PAYPAL_CLIENT_SECRET'),
-            'redirect'      => env(key: 'AUTHKIT_PAYPAL_REDIRECT'),
-            'sandbox_mode'  => (bool) env(key: 'AUTHKIT_PAYPAL_SANDBOX_MODE', default: true),
-            'scopes'        => ['openid', 'profile', 'email'],
-        ],
-    ],
 ];